requirements.txt
absl-py==2.3.1
academictorrents==2.3.3
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.4.0
bencode.py==2.0.0
bitstring==3.1.5
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
contourpy==1.3.2
cycler==0.12.1
diffusers==0.32.1
einops==0.8.1
fairscale==0.4.13
filelock==3.20.0
flatbuffers==25.12.19
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
future==0.16.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
grpcio==1.75.1
h5py==3.15.1
hf-xet==1.1.10
huggingface-hub==0.35.3
idna==3.11
imageio==2.37.0
importlib_metadata==8.7.0
Jinja2==3.1.6
joblib==1.5.2
keras==3.12.1
kiwisolver==1.4.9
lazy_loader==0.4
libclang==18.1.1
lightning-utilities==0.15.2
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
multidict==6.7.0
namex==0.1.0
networkx==3.4.2
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
optree==0.19.0
packaging==25.0
pillow==11.3.0
platformdirs==4.5.0
propcache==0.4.1
protobuf==6.32.1
psutil==7.1.0
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.2.5
PyPubSub==3.3.0
python-dateutil==2.9.0.post0
pytorch-fid==0.3.0
pytorch-lightning==2.4.0
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
rich==14.3.3
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
sentry-sdk==2.42.1
simsimd==6.5.3
six==1.17.0
smmap==5.0.2
stringzilla==4.2.1
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.12
tokenizers==0.22.1
torch==2.2.0+cu121
torchdiffeq==0.2.5
torchmetrics==1.8.2
torchvision==0.17.0+cu121
tqdm==4.67.1
transformers==4.57.1
triton==2.2.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wandb==0.22.2
Werkzeug==3.1.3
wrapt==2.1.1
yarl==1.22.0
zipp==3.23.0
